refactor: move LSP status indicator out of output.go
The status bar indicator landed in output.go while it was being wired to
the Output panel, but it is status bar code, not panel code.

Also stop stat'ing the server binary on every status sync — which runs
on every keystroke. Its presence only decides the icon for a server that
has not launched yet.
